+++ src/app/page.module.scss
@@ -2,8 +2,7 @@
 
 .main {
   display: flex;
-  width: 100vw;
-  height: 100vh;
+  min-height: 100vh;
 }
 
 .button {
