+++ src/app/table-menu/page.tsx
@@ -1,19 +1,13 @@
-import Header from "@/src/componets/Header";
+import Header from "@/src/components/Header";
 import styles from './TableMenu.module.scss';
-import { GetStaticProps, NextPage } from "next";
-import { ProductType, fetchProducts } from "@/src/services/products";
+import { NextPage } from "next";
 import { ReactNode } from "react";
-import ListProducts from "@/src/componets/ListProducts";
-
-export const getStaticProps: GetStaticProps = async () => {
-    const products = await fetchProducts()
-
-    return {props: {products}}
-}
+import ListProducts from "@/src/components/ListProducts";
+import dishes from '../../../database.json';
+import Footer from "@/src/components/common/Footer";
 
 const TableMenu: NextPage = (props: {
     children?: ReactNode
-    products?: ProductType[]
 }) =>{
     return (
         <>
@@ -22,9 +16,10 @@ const TableMenu: NextPage = (props: {
                 <h2>As melhores opções para o seu jantar!</h2>
                 <p>Navegue pelo nosso menu e descubra o melhor sabor para hoje!</p>
                 <div className={styles.menu}>
-                    {<ListProducts products={props.products!}/>}
+                    {<ListProducts products={dishes}/>}
                 </div>
             </main>
+            <Footer/>
         </>
     )
 }
