+++ src/components/CardProduct/index.module.scss
@@ -1,0 +1,2 @@
+@import '../../styles/colors';
+
