+++ src/components/CardProduct/index.tsx
@@ -1,6 +1,6 @@
-import Image from "next/image"
-import { useCommand } from "../hooks/useMenu"
-import { ProductType } from "../services/products"
+'use client'
+import { useCommand } from "../../hooks/useMenu"
+import { ProductType } from "../../services/products"
 import React from "react"
 
 
@@ -10,15 +10,17 @@ type CardProductsProps = {
 
 const CardProduct: React.FC<CardProductsProps> = ({ product }) =>{
     const { name, details, price, imageUrl } = product
-    const {addProduct} = useCommand()
+    const { addProduct } = useCommand()
 
     return(
         <div>
-            <Image src={imageUrl} alt={name} width={180} height={180}/>
+            <img src={imageUrl} alt={name} width={180} height={180}/>
             <h3>{name}</h3>
             <p>{details}</p>
             <p>{price}</p>
-            <button>Adicionar</button>
+            <button onClick={ () => {
+                addProduct(product)
+            }}>Adicionar</button>
         </div>
     )
 }
