+++ src/components/common/Footer/styles.module.scss
@@ -1,4 +1,4 @@
-@import '../styles/colors';
+@import '@/src/styles/_colors.scss';
 
 .footer{
     font-size: 14px;
