+++ src/services/products.ts
@@ -8,6 +8,6 @@ export type ProductType = {
 }
 
 export const fetchProducts = async () => {
-    const products: ProductType[] = await fetch(`${process.env.NEXT_PUBLIC_APIURL}/api/products`).then(res => res.json())
+    const products: ProductType[] = await fetch(`${process.env.NEXT_PUBLIC_URL}/api/products`).then(res => res.json())
     return products
 }
